Add admin access guard and store user roles
