+++ src/App.tsx
@@ -6,6 +6,7 @@ import {
   Outlet,
 } from "react-router-dom";
 import Layout from "./components/Layout";
+import RequireAdmin from "./components/RequireAdmin";
 import Login from "./pages/Login";
 import Register from "./pages/Register";
 import Dashboard from "./pages/Dashboard";
@@ -13,7 +14,6 @@ import PatientForm from "./pages/PatientForm";
 import PatientDetail from "./pages/PatientDetail";
 import PrintPreview from "./pages/PrintPreview";
 import AdminPanel from "./pages/AdminPanel";
-import { getFirestore, getDoc, doc } from "firebase/firestore";
 import { getAuth, onAuthStateChanged } from "firebase/auth";
 import { useEffect, useState } from "react";
 
@@ -39,36 +39,6 @@ function RequireAuth() {
   return <Outlet />;
 }
 
-function RequireAdmin() {
-  const [isAdmin, setIsAdmin] = useState(false);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const auth = getAuth();
-    const unsubscribe = onAuthStateChanged(auth, async (firebaseUser) => {
-      if (firebaseUser) {
-        const db = getFirestore();
-        const userDoc = await getDoc(doc(db, "users", firebaseUser.uid));
-        const data = userDoc.data();
-        setIsAdmin(data?.role === "admin");
-      } else {
-        setIsAdmin(false);
-      }
-      setLoading(false);
-    });
-    return () => unsubscribe();
-  }, []);
-
-  if (loading) {
-    return <div>Carregando...</div>;
-  }
-
-  if (!isAdmin) {
-    return <Navigate to="/" replace />;
-  }
-
-  return <Outlet />;
-}
 
 function App() {
   return (
